+++ packages/kilo-ui/src/components/icon-button.tsx
@@ -1,1 +1,29 @@
-export * from "@opencode-ai/ui/icon-button"
+import { Button as Kobalte } from "@kobalte/core/button"
+import { type ComponentProps, splitProps } from "solid-js"
+import { Icon, IconProps } from "./icon"
+
+export interface IconButtonProps extends ComponentProps<typeof Kobalte> {
+  icon: IconProps["name"]
+  size?: "small" | "normal" | "large"
+  iconSize?: IconProps["size"]
+  variant?: "primary" | "secondary" | "ghost"
+}
+
+export function IconButton(props: ComponentProps<"button"> & IconButtonProps) {
+  const [split, rest] = splitProps(props, ["variant", "size", "iconSize", "class", "classList"])
+  return (
+    <Kobalte
+      {...rest}
+      data-component="icon-button"
+      data-icon={props.icon}
+      data-size={split.size || "normal"}
+      data-variant={split.variant || "secondary"}
+      classList={{
+        ...split.classList,
+        [split.class ?? ""]: !!split.class,
+      }}
+    >
+      <Icon name={props.icon} size={split.iconSize ?? (split.size === "large" ? "normal" : "small")} />
+    </Kobalte>
+  )
+}
