+++ packages/ui/src/i18n/da.ts
@@ -138,11 +138,6 @@ export const dict = {
   "ui.message.revertMessage": "Nulstil til dette punkt",
   "ui.message.copyResponse": "Kopier svar",
   "ui.message.copied": "Kopieret!",
-  // kilocode_change start - message feedback
-  "ui.message.feedback.helpful": "Dette var nyttigt",
-  "ui.message.feedback.notHelpful": "Dette var ikke nyttigt",
-  "ui.message.feedback.clearRating": "Ryd bedømmelse",
-  // kilocode_change end
   "ui.message.revert": "Revert to here",
   "ui.message.interrupted": "Afbrudt",
   "ui.message.queued": "I kø",
